backend/pyproject.toml
[tool.poetry]
name = "strawberry-beats-backend"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.56.2"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
broadcaster = "^0.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
